Add delete image files functionality and error page

+++ workshop_petstagram/workshop_petstagram/main/views/pet_photos.py
@@ -2,9 +2,13 @@ from django.shortcuts import render, redirect
 
 from workshop_petstagram.main.forms import EditPetPhotoForm, AddPetPhotoForm
 from workshop_petstagram.main.models import PetPhoto
+from workshop_petstagram.main.templatetags.profiles import has_profile
 
 
 def add_pet_photo(request):
+    if not has_profile():
+        return redirect('error page')
+
     if request.method == "POST":
         form = AddPetPhotoForm(request.POST, request.FILES)
         if form.is_valid():
@@ -21,12 +25,15 @@ def add_pet_photo(request):
 
 
 def edit_pet_photo(request, pk):
+    if not has_profile():
+        return redirect('error page')
+
     photo = PetPhoto.objects.get(pk=pk)
-    if request.method == 'POST':
+    if request.method == "POST":
         form = EditPetPhotoForm(request.POST, instance=photo)
         if form.is_valid():
             form.save()
-            return redirect('dashboard')
+            return redirect('pet photo details', photo.pk)
     else:
         form = EditPetPhotoForm(instance=photo)
 
@@ -39,12 +46,18 @@ def edit_pet_photo(request, pk):
 
 
 def delete_pet_photo(request, pk):
+    if not has_profile():
+        return redirect('error page')
+
     pet_photo = PetPhoto.objects.get(pk=pk)
     pet_photo.delete()
     return redirect('dashboard')
 
 
 def like_pet_photo(request, pk):
+    if not has_profile():
+        return redirect('error page')
+
     pet_photo = PetPhoto.objects.get(pk=pk)
     pet_photo.likes += 1
     pet_photo.save()
@@ -53,6 +66,9 @@ def like_pet_photo(request, pk):
 
 
 def show_pet_photo_details(request, pk):
+    if not has_profile():
+        return redirect('error page')
+
     pet_photo = PetPhoto.objects \
         .prefetch_related('tagged_pets') \
         .get(pk=pk)
